Разбил все по папкам: 1 папка на ДЗ. Сделал часть первого задания.
